+++ libs/logic-apps-shared/src/designer-client-services/lib/base/copilot/copilotWorkflowEditorPrompt.ts
