+++ libs/logic-apps-shared/src/designer-client-services/lib/base/copilotWorkflowEditor.ts
